backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
pyjwt==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.1
reportlab==4.2.2
aiosqlite==0.20.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
